deepen: absorb derive-stake into form4-field-normalizer
Replaces the shallow lib/ai/utils/derive-stake.ts module with private
inlining into lib/email/form4-field-normalizer.ts. The 235-line module
had two production callers — form4-field-normalizer.ts itself
(normalizeForm4Data) and lib/ai/parsers/response-parser.ts (Form 4
branch), both of which already imported NormalizedTransaction and
formatNumberWithCommas from form4-field-normalizer. The split forced
a documented circular import that only worked by accident of function-
declaration hoisting:

    // Circular import: derive-stake.ts imports formatNumberWithCommas
    // from this file. Safe only because formatNumberWithCommas is a
    // `function` declaration and gets hoisted. Do NOT convert it to
    // `const fn = () => ...` or the cycle crashes at module init.

The seam existed to make the five-tier newStake precedence unit-testable
in isolation, but the leverage of that split was thin: the same
five-tier logic is exercised end-to-end through the normalizer's
interface by __tests__/email/form4-field-normalizer.test.ts (37 tests)
and __tests__/ai/parsers/response-parser-normalization.test.ts (47
tests) — 84 tests already crossing the deepened module's seam. The
circular-import hazard the split created was live-fire technical debt
the interface earned no leverage for.

deriveNewStake and detectNewStakeNarrativeMismatch remain exported from
form4-field-normalizer.ts so response-parser.ts can continue to reach
them through the one canonical Form 4 module. Local helpers
(parseStakeNumber, sortTxsChronologically, isCommonStock,
isDirectOwnership, isAllDerivative) become private to the file; the
COMMON_STOCK_VARIANTS and DERIVATIVE_CODES tables move with them.
Type exports (DeriveNewStakeSource / DeriveNewStakeOptions /
DeriveNewStakeResult) are preserved because the response-parser branch
consumes DeriveNewStakeResult.source and llmLegacyRaw fields.

Same shape as ADR-0002 (Historical Context / Why It Matters Guard /
Ticker Grounding Guard / Numeric Grounding Guard / Enrichment Flags):
extracted-for-testability shallow module inlined into its live-path
caller, with the interface tests already anchored at the caller's seam.

Locality: one file now owns every Form 4 normalization concern — the
transaction canonicalization, the previous/new-stake derivation, the
narrative-mismatch guard, and the shared numeric formatters. A
maintainer changing the five-tier precedence reads one file, not two.

Leverage: response-parser.ts drops one import line (deriveNewStake
comes in with its Form 4 siblings rather than through a separate
import), and no caller sees the missing seam because both callers
already stood on form4-field-normalizer.

Tests deleted: none — derive-stake had no test file of its own.
Tests added: none — the 84 tests at the form4-field-normalizer +
response-parser interfaces already covered every tier of the
five-tier precedence.

Diff: 3 files changed, 1 file deleted (~235 LOC), circular import gone.
